debut clean

+++ READ_GRID/get_letters/get_letters.c
@@ -96,10 +96,10 @@ int label_image_dfs(unsigned char **img, int **labels,
                     size_t x_final = 0;
                     int white_min2 = -1;
                     size_t x_final2 = 0;
-                    for (size_t x = cells[label - 1].x_min; 
+                    for (int x = cells[label - 1].x_min; 
                         x < cells[label - 1].x_max; x++){
                         double sum = 0;
-                        for (size_t y = cells[label - 1].y_min + 2; y 
+                        for (int y = cells[label - 1].y_min + 2; y 
                             <= cells[label - 1].y_max - 2; y++){
                             sum += img[y][x];
                         }
@@ -132,9 +132,9 @@ int label_image_dfs(unsigned char **img, int **labels,
                         cells[label-1].y_min = cells[label - 2].y_min;
                         cells[label-1].y_max = cells[label - 2].y_max;
 
-                        for (size_t x = cells[label-1].x_min; 
+                        for (int x = cells[label-1].x_min; 
                             x < cells[label-1].x_max; x++){
-                            for (size_t y = cells[label-1].y_min; 
+                            for (int y = cells[label-1].y_min; 
                                 y < cells[label-1].y_max; y++){
                                 if (labels[y][x] == label - 1)
                                     labels[y][x] = label;
@@ -157,9 +157,9 @@ int label_image_dfs(unsigned char **img, int **labels,
                         cells[label-1].y_min = cells[label - 2].y_min;
                         cells[label-1].y_max = cells[label - 2].y_max;
 
-                        for (size_t x = cells[label-1].x_min; 
+                        for (int x = cells[label-1].x_min; 
                             x < cells[label-1].x_max; x++){
-                            for (size_t y = cells[label-1].y_min; 
+                            for (int y = cells[label-1].y_min; 
                                 y < cells[label-1].y_max; y++){
                                 if (labels[y][x] == label - 1)
                                     labels[y][x] = label;
@@ -198,12 +198,12 @@ int label_image_dfs(unsigned char **img, int **labels,
                 > 1.2)){
                     int white_min = -1;
                     size_t x_final = 0;
-                    size_t diff = cells[label - 1].x_max - cells[label - 1].x_min + 1 ;
-                    for (size_t x = cells[label - 1].x_min + diff / 3 + 2; 
+                    int diff = cells[label - 1].x_max - cells[label - 1].x_min + 1 ;
+                    for (int x = cells[label - 1].x_min + diff / 3 + 2; 
                         x <= cells[label - 1].x_max - diff / 3 - 2; x++){
 
                         double sum = 0;
-                        for (size_t y = cells[label - 1].y_min + 2; 
+                        for (int y = cells[label - 1].y_min + 2; 
                             y <= cells[label - 1].y_max - 2; y++){
                             sum += abs(255 - img[y][x]);
                         }
@@ -235,9 +235,9 @@ int label_image_dfs(unsigned char **img, int **labels,
                         cells[label-1].y_min = cells[label - 2].y_min;
                         cells[label-1].y_max = cells[label - 2].y_max;
 
-                        for (size_t x = cells[label-1].x_min; 
+                        for (int x = cells[label-1].x_min; 
                             x < cells[label-1].x_max; x++){
-                            for (size_t y = cells[label-1].y_min; 
+                            for (int y = cells[label-1].y_min; 
                                 y < cells[label-1].y_max; y++){
 
                                 if (labels[y][x] == label - 1)
@@ -305,7 +305,7 @@ int label_image_dfs(unsigned char **img, int **labels,
         i += s;
     }
 
-    for (size_t i = 0; i < label; i ++){
+    for (int i = 0; i < label; i ++){
         cells[i].label = i;
     }
 
@@ -609,8 +609,8 @@ void Remove_too_far_mediane(struct Cell** cells, size_t n,
         i--;
     }
 
-    unsigned long int sum = 0;
-    unsigned int med = centers[tab1[size1/2]].center_y;
+    long int sum = 0;
+    int med = centers[tab1[size1/2]].center_y;
     for (size_t i = 0; i < size1; i++){
         sum += abs(centers[tab1[i]].center_y - med);
     }
@@ -722,8 +722,8 @@ char Same_families(struct family f1, struct family f2){
     return 1;
 }
 void Remove_same_families(struct family** all_families, int n){
-    for (size_t i = 0; i < n; i++){
-        size_t j = 0;
+    for (int i = 0; i < n; i++){
+        int j = 0;
         while (j < n && (*all_families)[i].completed != -1){
             if (  i != j && (*all_families)[j].completed != -1 
             && Same_families((*all_families)[i], (*all_families)[j]) ){
@@ -736,7 +736,7 @@ void Remove_same_families(struct family** all_families, int n){
 
 char contains(struct family f, int elt){
     for (size_t i = 0; i < f.size; i++){
-	if (f.tab[i].ind == elt)
+	if ((int)f.tab[i].ind == elt)
 	    return 1;
     }
     return 0;
@@ -745,14 +745,14 @@ char contains(struct family f, int elt){
 char Add_next_element(struct family** all_families, 
     struct Dist_with **families, int n, char second_call, struct Cell* cells){
     char changed = 0;
-    for (size_t i = 0; i < n; i++){
+    for (int i = 0; i < n; i++){
         struct family fam = (*all_families)[i];
         if (fam.completed == 0){
             changed = 1;
             double min = -1;
             size_t index = 0;
-            for (size_t j = 0; j < fam.size; j++){ 
-                while (fam.tab[j].actual < n - 1 && 
+            for (int j = 0; j < fam.size; j++){ 
+                while ((int)fam.tab[j].actual < n - 1 && 
             (contains(fam, families[fam.tab[j].ind][fam.tab[j].actual].index)
             || 
             (second_call == 1 &&
